# openapi.yaml
openapi: 3.0.1
info:
  title: Axionik Chat Shop API
  version: "1.0"

servers:
  - url: https://www.axionikai.com

paths:
  /chat-checkout:
    get:
      operationId: startShoppingSession
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: intent
          in: query
          schema:
            type: string
        - name: color
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: budget
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    type: string
                  email:
                    type: string
                  userId:
                    type: string
                  filters:
                    type: object

  /shop:
    get:
      operationId: getProducts
      parameters:
        - name: session
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                  products:
                    type: array
                    items:
                      type: object

  /add-to-cart:
    post:
      operationId: addToCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session:
                  type: string
                productId:
                  type: string
                qty:
                  type: number
      responses:
        "200":
          description: Added to cart

  /checkout/razorpay:
    post:
      operationId: createRazorpayCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session:
                  type: string
      responses:
        "200":
          description: Order created

  /chat-track-order:
    get:
      operationId: trackOrder
      parameters:
        - name: orderId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status